Hacky update to work with new Disqus layout
This works, but:

1. I'm pulling in jQuery through Tampermonkey, and then also embedding
jQuery-min in the script, which is horrible.
2. The code for generating the checkboxes doesn't work correctly unless
it's run after the page has loaded *and rendered*, it seems. So, for
now, I've set up Tampermonkey to only run it when asked, via the context
menu. This means it doesn't even highlight comments until you turn it
on... poor UX.

TODO:

1. Ditch jQuery (both in the outer code and the sandboxed code) and just
use vanilla JS.
2. Make the script run once the page has loaded and rendered, rather
than needing to be run manually.